modify video interfaces

+++ idl/video.thrift
@@ -1,7 +1,7 @@
 namespace go video
 
 struct BaseResp {
-    1:i64 status_code
+    1:i32 status_code
     2:string status_message
     3:i64 service_time
 }
